Check dtypes before calling C extension
If a 16-bit float tensor on the CPU was passed as the input to
quantize_blockwise or the output buffer for dequantize_blockwise, the
code was previously passing its pointer to the
c[de]quantize_blockwise_cpu_fp32 method, silently casting it to a 32-bit
float* and resulting in segfaults.

A similar issue occurs if the absmax/code arguments to
dequantize_blockwise are float16s, resulting in illegal memory accesses
on the GPU.

This commit adds some simple dtype guards to ensure the tensors have the
expected type before passing them to the C extension.